refactor: remove dto module
Having a whole set of DTOs in addition to the business/domain objects
seems to be of little benefit at this point and results in duplicating a
lot of code.  From this point, DTOs are reserved to bespoke
request/response objects (like AttestationRequest) that have no business
functionality and are used only for passing information.  Everything
else belongs to the model module.